requirements.txt
# Web server
fastapi==0.105.0
uvicorn==0.24.0
starlette==0.27.0
python-multipart==0.0.6
httpx==0.25.2
pydantic==2.4.2

# Database
motor==3.3.1
pymongo==4.6.0

# Background tasks
celery==5.4.0
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pandas==2.1.3
sec-edgar-downloader==5.0.2

# LLM
openai==1.60.2 